refactor: admin과 user 에서 공통으로 쓰이는 클래스는 common으로 이동/분리
