+++ src/main/java/kr/tennispark/point/common/application/service/PointService.java
@@ -1,13 +1,12 @@
-package kr.tennispark.point.user.application.service;
+package kr.tennispark.point.common.application.service;
 
-import java.util.List;
 import kr.tennispark.members.common.domain.entity.Member;
 import kr.tennispark.point.common.domain.entity.Point;
 import kr.tennispark.point.common.domain.entity.PointHistory;
+import kr.tennispark.point.common.domain.entity.enums.PointReason;
+import kr.tennispark.point.user.application.exception.PointNegativeValueException;
 import kr.tennispark.point.user.infrastrurcture.repository.PointHistoryRepository;
 import kr.tennispark.point.user.infrastrurcture.repository.PointRepository;
-import kr.tennispark.point.user.presentation.dto.response.GetMemberPointHistoryResponse;
-import kr.tennispark.point.user.presentation.dto.response.GetMemberPointResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,18 +14,22 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class UserPointService {
+public class PointService {
 
     private final PointRepository pointRepository;
     private final PointHistoryRepository pointHistoryRepository;
 
-    public GetMemberPointResponse getMemberPoint(Member member) {
+    @Transactional
+    public void applyPoint(Member member, int points, PointReason reason, String detail) {
+        if (points < 0) {
+            throw new PointNegativeValueException();
+        }
         Point point = pointRepository.getByMemberId(member.getId());
-        return new GetMemberPointResponse(point.getTotalPoint());
+
+        int signedPoint = reason.isEarned() ? Math.abs(points) : -Math.abs(points);
+
+        point.updatePoint(signedPoint);
+        pointHistoryRepository.save(PointHistory.of(point, member, points, reason, detail));
     }
 
-    public GetMemberPointHistoryResponse getMemberPointHistory(Member member) {
-        List<PointHistory> histories = pointHistoryRepository.findAllByMember(member);
-        return GetMemberPointHistoryResponse.of(histories);
-    }
 }
