+++ src/main/java/kr/tennispark/point/common/infrastructure/ProductRepository.java
@@ -1,4 +1,4 @@
-package kr.tennispark.point.user.infrastrurcture.repository;
+package kr.tennispark.point.common.infrastructure;
 
 import kr.tennispark.point.common.domain.entity.Product;
 import kr.tennispark.point.common.domain.exception.NoSuchProductException;
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
     default Product getById(Long id) {
         return findById(id)
